+++ src/components/index.ts
@@ -3,3 +3,4 @@ export * from './common';
 export * from './main';
 export * from './modal';
 export * from './section';
+export * from './mangeUserBook';
