+++ src/components/mangeUserBook/ReturnBookModal/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { MdLocationOff, MdNotListedLocation, MdCameraAlt } from 'react-icons/md';
 
-import { Modal } from '../CommonModal';
-import { StatusModal } from '../StatusModal';
+import { Modal } from '../../modal/CommonModal';
+import { StatusModal } from '../../modal/StatusModal';
 
 import * as S from './styled';
