+++ src/components/modal/index.ts
@@ -1,5 +1,3 @@
 export * from './CommonModal';
 export * from './DetailModal';
 export * from './StatusModal';
-export * from './AddClubModal';
-export * from './ReturnBookModal';
